backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-generativeai==0.7.2
sqlalchemy==2.0.30
pandas==2.2.2
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
lxml==5.2.2
python-multipart==0.0.9
openpyxl==3.1.2